requirements.txt
Flask==3.0.0
Office365-REST-Python-Client==2.5.3
python-dotenv==1.0.0
